requirements.txt
boto3==1.26.115
kubernetes==26.1.0
PyYAML==6.0